// 项目/学习开源项目/权限管理系统/OverallAuth2.0/代码/前端代码/overall-auth.-open.-vue-open/src/model/publicModel.ts
//select下拉框对应模型
export interface LayuiSelectModel {

    /// <summary>
    /// 显示名称
    /// </summary>
    label: string;

    /// <summary>
    /// 对应值
    /// </summary>
    value: string;

    /// <summary>
    /// 是否禁用
    /// </summary>
    disabled: boolean
}